chore(common): include vite plugin inside the config file
